Add junit test for kafka events consumer
Use LightweightPluginDaemonTest class infrastructure to boot the
gerrit server and deploy the plugin. This test assumes that the full
version of Kafka server is up and running prior to test execution.

An alternative approach (not done in this change) is to set up
testcontainers library with provided KafkaContainer: [1] and start
dockerized Kafka instance from within the JUnit test's setUp method.

Test Plan:

1. Run Kafka Server:
  $ docker run -p 2181:2181 -p 9092:9092 -d -e ADVERTISED_HOST=127.0.1 -e ADVERTISED_PORT=9092 spotify/kafka
2. Run the test:
  $ bazel test plugins/kafka-events:kafka_events_tests

[1] https://github.com/testcontainers/testcontainers-java/blob/master\
/modules/kafka/src/main/java/org/testcontainers/containers/KafkaContainer.java

Change-Id: Ia61868acc50d47b742126208fd078076dc2c2a94