Allow to send message to Kafka asynchronously
Introduce a new configuration setting plugin.kafka-events.sendAsync
that allows to control the synchrony of the send() operation to a
Kafka producer.

The existing behaviour on stable-2.16 do not wait for the Kafka
broker to confirm that the message is sent to all the in-sync replicas.
By disabling the asynchronous send of message, the send() would wait
for the message to be sent  and return the boolean result
of the operation.

NOTE: The drawback of the enabling the sendAsync parameter is that
      the broker-api would only return the status of the successful
      invocation of the Produder.send() operation and not the actual
      ack received by the Broker at the successful replication to
      all the replicas.

Bug: Issue 12604
Change-Id: Iec5d1efb033d978bf12c47317895c68604dffecb