Fix kafka-events replay messages feature
When requesting to reset the offset and consume messages from the
beginning, the subscriber has to wait first for the assignment of
partitions.

Failing to do so will cause the subscriber to consumer zero records,
since no partitions have yet been assigned.

An explicit poll() call before the seekToBeginning() ensures that the
consumer heartbeat is sent to kafka and thus partitions assigned.

Bug: Issue 14136
Change-Id: Ibc6a66507ebfc9bb6c67df9e576114bed8973e74