Fix kafka-events replay messages feature
When requesting to reset the offset and consume messages from the
beginning, the subscriber has to wait first for the assignment of
partitions.

Failing to do so will cause the subscriber to consume zero records,
since no partitions have yet been assigned.

Make an explicit poll() call before the seekToBeginning() to ensure that
the consumer heartbeat is sent to kafka and thus a partition is
assigned.

Bug: Issue 14136
Change-Id: Ibc6a66507ebfc9bb6c67df9e576114bed8973e74